+++ app/src/main/java/com/anu/calculator/expressionparser/Exp.java
@@ -2,16 +2,11 @@ package com.anu.calculator.expressionparser;
 
 /**
  * Abstract class Exp to represent expressions
- * @author dongwookim
- * @author xuyangshen
- *
- * You are not required to implement any function inside this class.
- * Please do not change anything inside this class as well.
  */
 
 public abstract class Exp {
 
 	public abstract String show();
-	public abstract int evaluate();
+	public abstract double evaluate();
 
 }
